Factor all the duplicated validation methods into `BaseControl`.
